third_party/blink/web_tests/external/wpt/url/url-constructor.any.worker_include=javascript-expected.txt

This is a testharness.js-based test.
[PASS] Loading data…
[PASS] Parsing: <javascript:/example.com/> against <http://example.org/foo/bar>
[PASS] Parsing: <javascript:example.com/> against <http://example.org/foo/bar>
[PASS] Parsing: <javascript:/example.com/> without base
[PASS] Parsing: <javascript:example.com/> without base
[FAIL] Parsing: <javascript:/../> without base
  assert_equals: href expected "javascript:/" but got "javascript:/../"
[PASS] Parsing: <> without base
[FAIL] Parsing: <javascript://example.com:8080/pathname?search#hash> without base
  assert_equals: host expected "example.com:8080" but got ""
[FAIL] Parsing: <javascript:///test> without base
  assert_equals: pathname expected "/test" but got "///test"
[FAIL] Parsing: <javascript://test/a/../b> without base
  assert_equals: href expected "javascript://test/b" but got "javascript://test/a/../b"
[FAIL] Parsing: <javascript://:443> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <javascript://test:test> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
[FAIL] Parsing: <javascript://[:1]> without base
  assert_throws_js: function "function() {\n          new URL(expected.input, base);\n        }" did not throw
Harness: the test ran to completion.
